+++ libcst/_nodes/tests/base.py
@@ -11,10 +11,10 @@ from typing import Any, Callable, Iterable, List, Optional, Sequence, Type
 from unittest.mock import patch
 
 import libcst as cst
-from libcst._nodes.internal import CodegenState, CodeRange, visit_required
+from libcst._nodes.internal import CodegenState, visit_required
 from libcst._types import CSTNodeT
 from libcst._visitors import CSTTransformer, CSTVisitorT
-from libcst.metadata.position_provider import PositionProvider
+from libcst.metadata import CodeRange, PositionProvider
 from libcst.testing.utils import UnitTest
