+++ libcst/_nodes/tests/test_return.py
@@ -7,8 +7,9 @@
 from typing import Any
 
 import libcst as cst
-from libcst import CodeRange, parse_statement
+from libcst import parse_statement
 from libcst._nodes.tests.base import CSTNodeTest
+from libcst.metadata import CodeRange
 from libcst.testing.utils import data_provider
