+++ libcst/_nodes/tests/test_small_statement.py
@@ -7,8 +7,8 @@
 from typing import Any
 
 import libcst as cst
-from libcst import CodeRange
 from libcst._nodes.tests.base import CSTNodeTest
+from libcst.metadata import CodeRange
 from libcst.testing.utils import data_provider
