+++ libcst/_nodes/tests/test_yield.py
@@ -7,9 +7,10 @@
 from typing import Callable, Optional
 
 import libcst as cst
-from libcst import CodeRange, parse_statement
+from libcst import parse_statement
 from libcst._helpers import ensure_type
 from libcst._nodes.tests.base import CSTNodeTest
+from libcst.metadata import CodeRange
 from libcst.testing.utils import data_provider
